Makes RecordHandler a django model
